requirements.txt
easyocr>=1.7.1
langchain>=0.0.352
openai>=1.3.7
PyPDF2>=3.0.1
streamlit>=1.29.0
streamlit-pdf-viewer>=0.0.3
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.1
numpy>=1.24.3
Pillow>=10.0.0
torch>=2.1.0
transformers>=4.36.0
python-docx>=1.0.1
